TopicPerUser implementat i actualitzat a StoreInfoActivity

+++ app/src/main/java/com/example/admin/e_torn/services/StoreService.java
@@ -18,7 +18,7 @@ public interface StoreService {
     Call<Store> getStoreById(@Path("id") String id);
 
     @GET("/averageTime/{id}")
-    Call<Integer> getStoreAverageTime(@Path("id") String id);
+    Call<Float> getStoreAverageTime(@Path("id") String id);
 
     @PUT("/stores/{storeId}/users/{userId}")
     Call<PostUserAddResponse> addUserToStore(@Path("storeId") String storeId, @Path("userId") String userId);
